HW-Isolation: Implemented Entry interface
In an OpenPOWER based system, a user or an application (in the
error path if the hardware is faulty) can isolate hardware and
the respective isolated hardware part will be ignored to init
during the next boot of the host.

The implemented entry interface will be used to maintain the isolated
hardware information and please refer below link to get more about
Entry interface.

https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-dbus-interfaces/+/43532

Signed-off-by: Ramesh Iyyar <rameshi1@in.ibm.com>
Change-Id: Ie3f240e3f541c266d0d790082f4d3aaa34193e0f