event: Fix, Remove runtime deallocation watcher if the Host is Off
A runtime deallocation watcher (created when an OS is running to
create a hardware status event when a hardware is deallocated)
is still running when the Host is off. We need to remove it
when the Host is Off otherwise, the watcher creates an event
unnecessarily if the hardware is deconfigured in the ipling.

Signed-off-by: Ramesh Iyyar <rameshi1@in.ibm.com>
Change-Id: I97bbca4a4fcc5f69788963511bf1cd0e9f512f6b