Nag: parse unresolved PEL's with Deconfigured bit set and add to JSON
Parse through all the unresolved error log objects(PEL's) and
check if the "Deconfigured" is set to true in the PEL json of
so add the callouts, and service actions if any to the JSON file.

Ensure that if the associated error log object has been already
parsed in other parsers ignore parsing.

Tested:
[
  {
    "SERVICABLE_EVENT": {
      "CEC_ERROR_LOG": {
        "Callout Section": {
          "Callout Count": "1",
          "Callouts": [
            {
              "CCIN": "5C67",
              "FRU Type": "Normal Hardware FRU",
              "Location Code": "U78DA.ND0.WZS003T-P0-C15",
              "Part Number": "F210110",
              "Priority": "Medium Priority",
              "Serial Number": "            "
            }
          ]
        },
        "DATE_TIME": "04/11/2023 09:39:15",
        "ERR_PLID": "0x500056F4",
        "SRC": "BD13E511"
      },
      "RESOURCE_ACTIONS": {
        "CURRENT_STATE": "DECONFIGURED",
        "GARD_RECORD": true,
        "REASON_DESCRIPTION": "predictive",
        "TYPE": "POWER10 Core"
      }
    }
  }
]

Test to check if the pel is already parsed if it is ignored.

Apr 11 11:27:36 rain104bmc faultlog[6564]: Ignoring already parsed (1342199540)

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: I93d10286cffdd814d0a4071c11ba00df7511440c