nag: capture data from pdbg target if error object is deleted
For guard record created with an error there will be an associated
error object(PEL).

At present data from error object and guard record is captured,
in the nag Json file.

If the error object is deleted by the user the guard record data
is not captured. This will be inconsistent as we can see guard
record on the system but the same not getting captured in NAG
json file.

Modified to capture the default data for error log object when
deleted.

Tested:
    "SERVICABLE_EVENT": {
      "CEC_ERROR_LOG": [
        {
          "Callout Section": {
            "Callout Count": 0,
            "Callouts": {}
          },
          "DATE_TIME": "00/00/0000 00:00:00",
          "ERR_PLID": 0,
          "SRC": 0
        },
        {
          "RESOURCE_ACTIONS": {
            "CURRENT_STATE": "DECONFIGURED",
            "GARD_RECORD": true,
            "REASON_DESCRIPTION": "fatal",
            "TYPE": "POWER10 Core"
          }
        }
      ]
    }

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ie9e4e3cd9b1c8dc261ba826c9eca321dff5719dd