nag: check for ECO mode when logging core details in nag file
At present "Cahce-Only cores" are logged as "POWER10 core" in nag
JSON file.

Hardware deconfiguration page in GUI displays core type as
"Cache-Only core" if ECO mode for the target is enabled.

Modified to check ECO mode if target type is "fc" or "core"
and logging as "Cache-Only core" in nag file.

Tested:
  {
    "DECONFIGURED": {
      "CURRENT_STATE": "CONFIGURED",
      "PHYS_PATH": "physical:sys-0/node-0/proc-1/eq-2/fc-0",
      "PLID": 0,
      "REASON_DESCRIPTION": "manual",
      "TYPE": "Cache-Only Core"
    }
  }

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: I5bc35f656e3d64175b3d605419f93fad32501dde