nag: log deconfig reason in upper case
At present deconfig reason for guard record received from guard
library is in lower case.

The deconfig reason computed in the faultlog based on DeconfiguredByReason
enum is in upper case.

To keep it synch coverting the deconfig reason received for guard records
from guard library to uppercase.

Example:
“REASON_DESCRIPTION”: “manual”,
changed to
“REASON_DESCRIPTION”: “MANUAL”,

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: I001ed49c59f3218693bc73992024e0ca41ec9d9f