Revert "common: Consider ocmb as parent target instead of dimm"
This reverts commit 8b8286d81692b364da24334011388a40e0c2911a.

Location Code for OCMB is same as DIMM on rainier and everest, so we decided to
fetch it from OCMB target as it is available in device tree. But the Location
Code seems to be different for OCMB and DIMM on bonnell, so reverting this
commit as it is causing some parent DIMM numbering issues on bonnell.