faultlog - cater for spare core gard records
Do not notify the caller if there are only spare core
guard records, but collect spare core gard reocrd details
if there are other gard records or unresolved PELS.

DeconfiguredHW PEL is created if the gard record count
or unresloved PEL's count is greater than zero, so change
is made to not to consider spare core gard record as part
of calculating the gard record count.

HMC monitors DeconfiguredHw PEL and initiates NAGDUMP once
it detects the DeconfiguredHW PEL is created.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: I94419ab8f7343daf182c6a141b46d60608d8807e