Remove errorlog association on PEL deletion
If a system guard's associated PEL is removed, the update to the guard
record is skipped in the restore path.

As the guard still exists, it is better to update the guard without the
errorlog association. The gui could show a system guard (predictive/
fatal) without the errorlog.

Changes cherry-picked from 1110

Tested and found it working as expected

Signed-off-by: deepakalak <deepakala.karthikeyan@ibm.com>