Remove errorlog association on PEL deletion
If a system guard's associated PEL is removed, the guard was skipped to
be updated in the GUI as the association is lost. The following error
was displayed.

```Skipping to restore a given isolated hardware [23 01 00 02 00 03 00 ] :
Due to failure to get BMC error log path by isolated hardware EID (aka
PEL ID) [1342178283]```

As the guard still exists, it is better to update the guard without the
errorlog association. The gui could show a system guard (predictive/
fatal) without the errorlog.

Cherry picked the change from 1110.

Tested and found it working as expected

Signed-off-by: deepakalak <deepakala.karthikeyan@ibm.com>