Move the business logic to guard repo
The decision to clear an individual record, is based on multiple factors
including manual or system generated and what type of target is guarded.
And this decision shall be made at the guard module, so the behaviour is
consistent and the caller does not duplicate the logic

This commit removes the logic from this repo.

https://github.com/open-power/guard/pull/27/ moves the logic to guard
repo

Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>